Handle token attachment and detachment
If the USB token assigned to a container is attached and/or detached during runtime,
it is necessary to reflect this fact in the corresponding data structures of the scd and cmld.
Therefore, this commit handles the uevents triggered by token de- and attachments
and updates the token state in the scd and cmld.

+++ common/str.h
@@ -205,7 +205,7 @@ str_length(str_t *str);
  * @return The resulting hex dump
  */
 str_t *
-str_hexdump_new(unsigned char *mem, int len);
+str_hexdump_new(unsigned char *mem, size_t len);
 
 /**
  * Frees the allocated string memory.
